More changes including the move of setup_logger call to BMI initialize() function in bmi_cfe.c
